[FLINK-25266][tests] Consolidate Kafka testutils to make them part of the test jar
