+++ flink-connectors/flink-connector-kafka/src/test/java/org/apache/flink/connector/kafka/testutils/KafkaPartitionDataWriter.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.connector.kafka.source.testutils;
+package org.apache.flink.connector.kafka.testutils;
 
 import org.apache.flink.connectors.test.common.external.SourceSplitDataWriter;
